refactor: consolidate mock fixtures for Home Assistant and config entry in tests
