+++ custom_components/power_max_tracker/tests/test_sensor.py
@@ -8,10 +8,6 @@ import pytest
 from datetime import datetime
 from unittest.mock import MagicMock, patch, AsyncMock
 
-from homeassistant.core import HomeAssistant
-from homeassistant.config_entries import ConfigEntry
-
-from custom_components.power_max_tracker.coordinator import PowerMaxCoordinator
 from custom_components.power_max_tracker.sensor import (
     MaxPowerSensor,
     SourcePowerSensor,
@@ -27,38 +23,6 @@ from custom_components.power_max_tracker.const import (
     CONF_BINARY_SENSOR,
     DOMAIN,
 )
-
-
-@pytest.fixture
-def mock_hass():
-    """Create a mock Home Assistant instance."""
-    hass = MagicMock(spec=HomeAssistant)
-    hass.states = MagicMock()
-    hass.data = {}
-    hass.config = MagicMock()
-    hass.config.config_dir = "/tmp/test_hass_config"
-    hass.loop = MagicMock()
-    return hass
-
-
-@pytest.fixture
-def mock_config_entry():
-    """Create a mock config entry."""
-    entry = MagicMock(spec=ConfigEntry)
-    entry.entry_id = "test_entry_id"
-    entry.data = {
-        CONF_SOURCE_SENSOR: "sensor.test_power",
-        CONF_MONTHLY_RESET: False,
-        CONF_NUM_MAX_VALUES: 2,
-        CONF_BINARY_SENSOR: None,
-    }
-    return entry
-
-
-@pytest.fixture
-def coordinator(mock_hass, mock_config_entry):
-    """Create a PowerMaxCoordinator instance."""
-    return PowerMaxCoordinator(mock_hass, mock_config_entry)
 
 
 class TestMaxPowerSensor:
